Split the viewer into multiple components
Each "panel" (media selector, media view, media info) are now resource
page blocks that can be placed anywhere on the page.
They communicate with each other so that selecting a media in the media
selector will update other parts present on the page.
The full viewer is still available.

There is a new "controls" element below the media view. The fullscreen
button and the download link from the media info panel moved to this
location.