WIP: Break the viewer into multiple parts
Each "panel" (media selector, media view, media info) are now resource
page blocks that can be placed anywhere on the page.
They communicate with each other so that selecting a media in the media
selector will update other parts present on the page.
The full viewer is still available.

To be done:
- Finish the "controls" component inside media-view
  - Pass all parameters to the fullscreen viewer so it behave exactly
    like the full viewer
  - Move the download button into this component ?