Add navigation buttons (previous and next) inside the main view controls
This needed some kind of shared state between the different JS
components, that's what the new class OctopusViewer is. Every component
references an OctopusViewer instance. Some may share the same, some may
have a unique instance. This allows to have several viewer "parts" on
the same page communicating with each other (or not).

All components have an optional attribute "instance-key" that can be
used to connect components together (by using the same instance key) or
to make sure the component is using its own unique instance (by using a
unique instance key).
By default this is not needed as all components that have the same
"site-slug" and "media-query" attributes share the same instance.